requirements.txt
coreapi==2.3.3
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.1
django-cors-headers==4.1.0
django-filter==23.2
django-ratelimit==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-extensions==0.7.1
drf-yasg==1.21.6
graphene-django==3.1.2
gunicorn==20.1.0
jsonfield==3.1.0
lxml==4.9.2
matplotlib==3.7.1
pandas==1.2.0
psycopg2-binary==2.9.6
Pygments==2.15.1
ruamel.yaml==0.17.32
scikit-learn==1.2.2
scipy==1.11.0
statsmodels==0.14.0
uvicorn==0.13.3
whitenoise==6.5.0
yahoo-finance-api2==0.0.12
django-tidb==3.0.0
python-decouple
